requirements.txt
Flask
mysql-connector-python
Flask-MySQL
werkzeug
cloudinary
gunicorn
sqlalchemy
pandas
secure-smtplib
uuid
datetime
timedelta
PyJWT